[move] Initial trace format and implementation (#18729)
This PR adds the initial trace format, and the implementation of this
trace format in the VM. I've gone through the tracing format with most
of you synchronously so won't write out all the details here (but I will
take it on to write a spec for the trace format once this is done so
other folks can use it when consuming this format). Happy to go through
it at any point in real time.

Other TODO: right now the `MoveValue`s only support serialize and not
deserialize back into Rust so we can only push a trace out from Rust but
not consume it. The next thing I'm working on right now is adding
support for that, and it may be either an update to this PR, or an
additional PR depending on how involved that is...

Tests and integration of this into the Move CLI (not the Sui CLI yet) is
in the PR above this one.

This keeps the tracing largely feature-gated in the VM, and the only
additional overhead/change at runtime with `gas-profiling` turned off is
the additional argument, but this argument is unused and should be
optimized away (and at worst only add essentially no overhead).

I kept the `gas-profiling` feature flag and gated the new tracing under
it. The plan being to eventually rename that flag at the same time that
we update test coverage and gas profiling to use this new tracing format
as well (but that's for a couple future PRs!).

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: